One Front Door/SourceCode/MerchAngular2ReleaseiPhone/src/app/common/storelist/storelist.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { StorelistComponent } from './storelist.component';

describe('Component: Storelist', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [StorelistComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([StorelistComponent],
      (component: StorelistComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(StorelistComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(StorelistComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-storelist></app-storelist>
  `,
  directives: [StorelistComponent]
})
class StorelistComponentTestController {
}

